+++ visivo/models/project.py
@@ -3,6 +3,9 @@ from typing import Any, List, Optional, Union
 from visivo.models.alert import AlertField, Alert
 
 from visivo.models.include import Include
+from visivo.models.model import Model
+from visivo.models.models.fields import ModelField
+from visivo.models.models.sql_model import SqlModel
 
 
 from .base.parent_model import ParentModel
@@ -11,7 +14,6 @@ from .chart import Chart
 from .trace import Trace
 from .target import Target
 from .table import Table
-from .model import ModelField, Model, SqlModel
 from .defaults import Defaults
 from typing import List
 from .base.named_model import NamedModel
