+++ visivo/query/jobs/run_trace_job.py
@@ -1,9 +1,8 @@
-from visivo.logging.logger import Logger
 from visivo.models.base.parent_model import ParentModel
-from visivo.models.model import CsvScriptModel, Model
+from visivo.models.model import Model
+from visivo.models.models.csv_script_model import CsvScriptModel
 from visivo.models.project import Project
 from visivo.models.target import Target
-from visivo.models.trace import Trace
 from visivo.query.aggregator import Aggregator
 from visivo.query.jobs.job import (
     Job,
